+++ mobile/android/android-components/components/browser/toolbar/src/main/java/mozilla/components/browser/toolbar/BrowserToolbar.kt
@@ -333,13 +333,23 @@ class BrowserToolbar @JvmOverloads constructor(
 
     /**
      * Switches to URL editing mode.
+     *
+     * @param cursorPlacement Where the cursor should be placed after focusing on the URL input field.
      */
-    override fun editMode() {
+    override fun editMode(cursorPlacement: Toolbar.CursorPlacement) {
         val urlValue = if (searchTerms.isEmpty()) url else searchTerms
         edit.updateUrl(urlValue.toString(), false)
         updateState(State.EDIT)
         edit.focus()
-        edit.selectAll()
+
+        when (cursorPlacement) {
+            Toolbar.CursorPlacement.ALL -> {
+                edit.selectAll()
+            }
+            Toolbar.CursorPlacement.END -> {
+                edit.selectEnd()
+            }
+        }
     }
 
     /**