+++ mobile/android/android-components/components/concept/toolbar/src/main/java/mozilla/components/concept/toolbar/Toolbar.kt
@@ -183,9 +183,12 @@ interface Toolbar {
     fun displayMode()
 
     /**
-     * Switches to URL editing mode (from displaying mode) if supported by the toolbar implementation.
+     * Switches to URL editing mode (from display mode) if supported by the toolbar implementation,
+     * and focuses the URL input field based on the cursor selection.
+     *
+     * @param cursorPlacement Where the cursor should be set after focusing on the URL input field.
      */
-    fun editMode()
+    fun editMode(cursorPlacement: CursorPlacement = CursorPlacement.ALL)
 
     /**
      * Dismisses the display toolbar popup menu
@@ -492,6 +495,21 @@ interface Toolbar {
          * The site does not show a dot indicator.
          */
         NONE,
+    }
+
+    /**
+     * Indicates where the cursor should be set after focusing on the URL input field.
+     */
+    enum class CursorPlacement {
+        /**
+         * All of the text in the input field should be selected.
+         */
+        ALL,
+
+        /**
+         * No text should be selected and the cursor should be placed at the end of the text.
+         */
+        END,
     }
 }