+++ mobile/android/android-components/components/feature/toolbar/src/test/java/mozilla/components/feature/toolbar/ToolbarAutocompleteFeatureTest.kt
@@ -100,7 +100,7 @@ class ToolbarAutocompleteFeatureTest {
             fail()
         }
 
-        override fun editMode() {
+        override fun editMode(cursorPlacement: Toolbar.CursorPlacement) {
             fail()
         }