+++ mobile/android/android-components/components/feature/toolbar/src/test/java/mozilla/components/feature/toolbar/ToolbarInteractorTest.kt
@@ -85,7 +85,7 @@ class ToolbarInteractorTest {
             fail()
         }
 
-        override fun editMode() {
+        override fun editMode(cursorPlacement: Toolbar.CursorPlacement) {
             fail()
         }